More moving things around and cleaning up
