+++ splitgraph/core/_common.py
@@ -45,7 +45,7 @@ def manage_audit_triggers():
     """
 
     engine = get_engine()
-    from splitgraph.core.repository import get_current_repositories
+    from splitgraph.core.engine import get_current_repositories
     repos_tables = [(r.to_schema(), t) for r, head in get_current_repositories() if head is not None
                     for t in set(engine.get_all_tables(r.to_schema())) & set(r.get_image(head).get_tables())]
     tracked_tables = engine.get_tracked_tables()
