+++ splitgraph/splitfile/execution.py
@@ -10,7 +10,8 @@ from random import getrandbits
 from splitgraph._data.provenance import store_import_provenance, store_sql_provenance, store_mount_provenance, \
     store_from_provenance
 from splitgraph.config import CONFIG
-from splitgraph.core.repository import to_repository, Repository, local_clone, clone, repository_exists, lookup_repo
+from splitgraph.core.engine import repository_exists, lookup_repo
+from splitgraph.core.repository import to_repository, Repository, local_clone, clone
 from splitgraph.engine import get_engine
 from splitgraph.exceptions import SplitGraphException
 from splitgraph.hooks.mount_handlers import get_mount_handler
