+++ test/splitgraph/conftest.py
@@ -5,8 +5,10 @@ import pytest
 from splitgraph._data.common import ensure_metadata_schema
 from splitgraph._data.registry import _ensure_registry_schema, unpublish_repository, setup_registry_mode, \
     toggle_registry_rls
-from splitgraph.core.repository import to_repository as R, Repository, cleanup_objects, mount, get_current_repositories
+from splitgraph.core.engine import cleanup_objects, get_current_repositories
+from splitgraph.core.repository import to_repository as R, Repository
 from splitgraph.engine import get_engine, ResultShape, switch_engine
+from splitgraph.hooks.mount_handlers import mount
 
 PG_MNT = R('test/pg_mount')
 PG_MNT_PULL = R('test_pg_mount_pull')
